fix(rsvp): email non-members promoted by member RSVP changes (Issue 629)
email_promoted_non_members only ran from the public RSVP path, so a
non-member promoted off the waitlist because a member left or changed
their RSVP was never emailed — and can't see the in-app notification,
leaving them with no signal a spot opened.

Extract the shared RSVP email helpers into community/_rsvp_emails.py (a
new module, since _public_rsvp imports from _event_rsvps so the reverse
would be circular, and _event_helpers is already near the size cap), and
call email_promoted_non_members from upsert_rsvp and delete_rsvp after
commit.
