+++ backend/community/_event_rsvps.py
@@ -35,6 +35,7 @@ from community._event_schemas import (
     TextRecipientsOut,
 )
 from community._events import _can_edit_event, _enforce_event_read_visibility
+from community._rsvp_emails import email_promoted_non_members
 from community._shared import ErrorOut
 from community._validation import Code, raise_validation
 from community.models import Event, EventRSVP, RSVPStatus
@@ -170,7 +171,7 @@ def upsert_rsvp(request, event_id: UUID, payload: RSVPIn):
     _validate_rsvp_status(payload.status)
 
     with transaction.atomic():
-        final_status, _promoted_user_ids = _apply_rsvp_in_transaction(
+        final_status, promoted_user_ids = _apply_rsvp_in_transaction(
             event_id, request.auth, payload.status, payload.has_plus_one
         )
 
@@ -187,6 +188,7 @@ def upsert_rsvp(request, event_id: UUID, payload: RSVPIn):
         raise_validation(Code.Event.NOT_FOUND, status_code=404)
     # Actor already has the fresh event in this response, so exclude them.
     broadcast_capacity_change(event_id, exclude_user_ids={str(request.auth.pk)})
+    email_promoted_non_members(request, event, promoted_user_ids)
     return Status(200, _event_out(event, request.auth))
 
 
@@ -342,9 +344,10 @@ def delete_rsvp(request, event_id: UUID):
             raise_validation(Code.Event.RSVPS_CLOSED_PAST, status_code=400)
         was_attending = rsvp.status == RSVPStatus.ATTENDING
         rsvp.delete()
+        promoted_user_ids = promote_from_waitlist(event) if was_attending else []
         if was_attending:
-            promote_from_waitlist(event)
             broadcast_capacity_change(event_id, exclude_user_ids={str(request.auth.pk)})
 
     audit_log(logging.INFO, "rsvp_deleted", request, target_type="event", target_id=str(event_id))
+    email_promoted_non_members(request, event, promoted_user_ids)
     return Status(204, None)
