+++ backend/community/_public_rsvp.py
@@ -2,17 +2,9 @@ import logging
 
 from config.audit import audit_log
 from config.ratelimit import client_ip, rate_limit
-from django.conf import settings
 from django.db import IntegrityError, transaction
-from django.utils import timezone
 from ninja import Router
 from ninja.responses import Status
-from notifications._email_helpers import (
-    RsvpEmailDetails,
-    send_rsvp_confirmation_email,
-    send_rsvp_waitlist_promoted_email,
-)
-from notifications.email_sender import get_email_sender
 from pydantic import BaseModel, EmailStr, Field
 from users.models import NonMemberRsvpToken, User
 
@@ -20,7 +12,8 @@ from community._event_helpers import _event_out, broadcast_capacity_change
 from community._event_rsvps import _apply_rsvp_in_transaction, _validate_rsvp_status
 from community._event_schemas import EventOut
 from community._field_limits import FieldLimit
-from community._shared import ErrorOut, _validate_phone, logger, validate_display_name
+from community._rsvp_emails import email_promoted_non_members, send_confirmation_email
+from community._shared import ErrorOut, _validate_phone, validate_display_name
 from community._validation import Code, raise_validation
 from community.models import Event, EventStatus, EventType, PageVisibility, RSVPStatus
 
@@ -142,80 +135,6 @@ def _resolve_non_member(*, request, name: str, email: str, phone: str) -> User:
     return _create_non_member(name, email, phone)
 
 
-def _format_event_when(event: Event) -> str:
-    if event.datetime_tbd or event.start_datetime is None:
-        return "to be decided"
-    local = timezone.localtime(event.start_datetime)
-    return local.strftime("%A, %B %d at %I:%M %p").replace(" 0", " ")
-
-
-def _event_links(event: Event) -> list[str]:
-    return [link for link in (event.whatsapp_link, event.partiful_link, event.other_link) if link]
-
-
-def _email_details(event: Event, user: User, token_str: str) -> RsvpEmailDetails:
-    return RsvpEmailDetails(
-        to=user.email,
-        display_name=user.display_name,
-        event_title=event.title,
-        event_when=_format_event_when(event),
-        event_location=event.location,
-        event_links=_event_links(event),
-        manage_url=f"{settings.FRONTEND_BASE_URL}/my-rsvps?token={token_str}",
-        join_url=f"{settings.FRONTEND_BASE_URL}/join",
-    )
-
-
-def _log_email_failure(request, event: Event, user: User, exc: Exception) -> None:
-    logger.warning("public rsvp email failed", exc_info=True)
-    audit_log(
-        logging.WARNING,
-        "public_rsvp_email_failed",
-        request,
-        target_type="event",
-        target_id=str(event.id),
-        details={"user_id": str(user.pk), "error": str(exc)},
-    )
-
-
-def _send_confirmation_email(
-    request, event: Event, user: User, token_str: str, waitlisted: bool
-) -> None:
-    """Best-effort confirmation email. A send failure must NOT roll back the RSVP."""
-    if not user.email:
-        return
-    try:
-        result = send_rsvp_confirmation_email(
-            sender=get_email_sender(),
-            details=_email_details(event, user, token_str),
-            waitlisted=waitlisted,
-        )
-        if not result.success:
-            raise RuntimeError(result.error or "send returned failure")
-    except Exception as exc:
-        _log_email_failure(request, event, user, exc)
-
-
-def _email_promoted_non_members(request, event: Event, promoted_user_ids: list[str]) -> None:
-    """Email any promoted non-members a fresh manage link. Best-effort per user."""
-    if not promoted_user_ids:
-        return
-    promoted = User.objects.filter(id__in=promoted_user_ids, is_member=False, email__isnull=False)
-    for user in promoted:
-        if not user.email:
-            continue
-        try:
-            token = NonMemberRsvpToken.issue_or_extend(user)
-            result = send_rsvp_waitlist_promoted_email(
-                sender=get_email_sender(),
-                details=_email_details(event, user, token.token),
-            )
-            if not result.success:
-                raise RuntimeError(result.error or "send returned failure")
-        except Exception as exc:
-            _log_email_failure(request, event, user, exc)
-
-
 @router.post(
     "/public/events/{event_id}/rsvp/",
     response={200: PublicRsvpOut, 400: ErrorOut, 404: ErrorOut, 409: ErrorOut, 429: ErrorOut},
@@ -261,8 +180,8 @@ def submit_public_rsvp(request, event_id, payload: PublicRsvpIn):
     )
 
     waitlisted = final_status == RSVPStatus.WAITLISTED
-    _send_confirmation_email(request, event, user, token.token, waitlisted)
-    _email_promoted_non_members(request, event, promoted_user_ids)
+    send_confirmation_email(request, event, user, token.token, waitlisted)
+    email_promoted_non_members(request, event, promoted_user_ids)
 
     fresh_event = (
         Event.objects.select_related("created_by")
